# app/ngx-admin/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ngx-admin Backend Bundle API",
    "description": "ngx-admin Backend Bundle API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Bearer <my own JWT token>"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Orders list",
        "description": "Gets orders list",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "filterByname",
            "in": "query",
            "description": "Should filter by name field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBytype",
            "in": "query",
            "description": "Should filter by type field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBystatus",
            "in": "query",
            "description": "Should filter by status name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBycountry",
            "in": "query",
            "description": "Should filter by country",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBysum",
            "in": "query",
            "description": "Should filter by sum",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filterBydate",
            "in": "query",
            "description": "Should filter by date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Should sort by field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Order direction ASC / DESC",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of orders (filtered orders by query params) and total count of orders in response",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "$ref": "#/definitions/Order"
                  }
                },
                "totalCount": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Add new order",
        "description": "Add new order",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "OrderDto",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/OrderDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get new order",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Order",
        "description": "Gets order",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Should return order",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Edit order",
        "description": "Edit order",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "OrderDto",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Should return edited user",
            "schema": {
              "$ref": "#/definitions/OrderDto"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Delete order",
        "description": "Delete order",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete order",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/orders-aggregated": {
      "get": {
        "tags": [
          "Orders Aggregated"
        ],
        "summary": "Orders list",
        "description": "Gets orders list",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "aggregation",
            "in": "query",
            "description": "Should filter by period of time. Default value : year",
            "schema": {
              "type": "string",
              "default": "year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of orders",
            "schema": {
              "type": "object",
              "properties": {
                "axisXLabels": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "chartLabel": {
                  "type": "string"
                },
                "lines": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "values": {
                        "type": "array",
                        "items": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders-aggregated/profit": {
      "get": {
        "tags": [
          "Orders Aggregated"
        ],
        "summary": "Profit list data",
        "description": "Gets profit list data",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "aggregation",
            "in": "query",
            "description": "Should filter by period of time. Default value : year",
            "schema": {
              "type": "string",
              "default": "year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of aggregated profit data",
            "schema": {
              "type": "object",
              "properties": {
                "axisXLabels": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "chartLabel": {
                  "type": "string"
                },
                "lines": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "values": {
                        "type": "array",
                        "items": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders-aggregated/country": {
      "get": {
        "tags": [
          "Orders Aggregated"
        ],
        "summary": "Country list data",
        "description": "Gets country list data",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "countryCode",
            "in": "query",
            "description": "Should filter by country code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of aggregated country",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "orderType": {
                    "type": "string"
                  },
                  "count": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orders-aggregated/summary": {
      "get": {
        "tags": [
          "Orders Aggregated"
        ],
        "summary": "Orders summary count",
        "description": "Gets summary count of orders",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of aggregated country",
            "schema": {
              "type": "object",
              "properties": {
                "today": {
                  "type": "number"
                },
                "lastMonth": {
                  "type": "number"
                },
                "lastWeek": {
                  "type": "number"
                },
                "marketplace": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "sum": {
          "type": "object",
          "properties": {
            "currency": {
              "type": "string"
            },
            "value": {
              "type": "number"
            }
          }
        },
        "country": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number"
            },
            "name": {
              "type": "string"
            }
          }
        }
      }
    },
    "OrderDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "sum": {
          "type": "object",
          "properties": {
            "currency": {
              "type": "string"
            },
            "value": {
              "type": "number"
            }
          }
        },
        "country": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number"
            },
            "name": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}